Testes de nova funcionalidade

+++ src/main/java/com/kuuhaku/command/commands/information/ProfileCommand.java
@@ -56,10 +56,10 @@ public class ProfileCommand extends Command {
 	@Override
 	public void execute(User author, Member member, String rawCmd, String[] args, Message message, MessageChannel channel, Guild guild, String prefix) {
 		Account acc = AccountDAO.getAccount(author.getId());
+		com.kuuhaku.model.persistent.Member mb = MemberDAO.getMemberById(author.getId() + guild.getId());
 		channel.sendMessage(ShiroInfo.getLocale(I18n.PT).getString("str_generating-profile")).queue(m -> {
 			try {
-				System.out.println(Helper.getFileType(MemberDAO.getMemberById(author.getId() + guild.getId()).getBg()));
-				if (acc.hasAnimatedBg()) {
+				if (acc.hasAnimatedBg() && Helper.getFileType(mb.getBg()).contains("gif")) {
 					File pf = Profile.applyAnimatedBackground(MemberDAO.getMemberById(author.getId() + guild.getId()), Profile.makeProfile(member, guild));
 					channel.sendMessage(MessageFormat.format(ShiroInfo.getLocale(I18n.PT).getString("str_profile"), author.getAsMention())).addFile(pf, "perfil.gif").queue(s -> m.delete().queue());
 				} else
